requirements.txt
# --- Core Framework ---
fastapi==0.115.0
uvicorn==0.30.6

# --- Database ---
pymongo==4.8.0

# --- Authentication ---
python-jose==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.12
pyjwt==2.9.0

# --- Optional: For email validation (used by EmailStr)
email-validator==2.2.0

pytest==8.3.3 
pytest-asyncio==0.24.0 
httpx==0.27.2